Launch multiple kernels for large batch sizes (> 65535) & Use int64 index (#4)
* rename test files

* use int64 index

* fix single precision

* support large batch size

+++ actnn/actnn/cpp_extension/backward_func.cc
@@ -17,7 +17,7 @@ std::tuple<Tensor, Tensor, Tensor, int64_t, int64_t> prepare_layer_norm_inputs(
     const Tensor& weight /* optional */,
     const Tensor& bias /* optional */) {
 
-  const int normalized_ndim = normalized_shape.size();
+  const int64_t normalized_ndim = normalized_shape.size();
   TORCH_CHECK(
       normalized_ndim >= 1,
       "Expected normalized_shape to be at least 1-dimensional, i.e., ",
